MAINT: Cleanup and use is_matrix_loaded more
