+++ src/helpers.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+bool is_matrix_loaded();
